requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via httpx
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
exceptiongroup==1.2.2
    # via anyio
gigachat==0.1.35
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via gigachat
idna==3.10
    # via
    #   anyio
    #   httpx
pydantic==2.9.2
    # via gigachat
pydantic-core==2.23.4
    # via pydantic
python-dotenv==1.0.1
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
typing-extensions==4.12.2
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
